refactor for clean
